devjobscout/pyproject.toml
[project]
name = "devjobscout"
version = "0.1.0"
description = "LinkedIn job scraper with browser-use and AI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "browser-use[langchain]>=0.9.5",
    "langchain-ollama>=1.0.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.31.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "python-telegram-bot>=20.7",
    "notion-client>=2.2.0",
    "playwright>=1.40.0",
    "pypdf>=3.17.0",
    "python-docx>=1.1.0",
]
